+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#76] get & set backup-schedule
 - [#76] get retention-policy
 - [#77] Create backups in ui
+- [#80] Create restores in ui
 
 ## [v1.7.3] - 2025-10-13
 ### Changed